+++ src/main/java/com/example/Biblioteca/service/CategoriaService.java
@@ -5,8 +5,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.Biblioteca.Excepciones.ValidacionException;
 import com.example.Biblioteca.modelo.Categoria;
 import com.example.Biblioteca.repository.CategoriaRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class CategoriaService implements IGenericService<Categoria>{
@@ -16,12 +19,19 @@ public class CategoriaService implements IGenericService<Categoria>{
 
     @Override
     public void guardar(Categoria t) {
+        if(!catRepo.existsByNombre(t.getNombre())){     // si se agrega otra validación, crear una clase especial para validar categoría
         catRepo.save(t);
+        }else{
+            throw new ValidacionException("Nombre", "Existe ya un registro de esta categoría");
+        }
     }
 
     @Override
     public Categoria obtenerUno(Long id) {
-       return catRepo.findById(id).orElse(null);
+        if(id == null){
+            throw new IllegalArgumentException("Debe ingresar id");
+        }
+       return catRepo.findById(id).orElseThrow(() -> new EntityNotFoundException("Categoría no encontrada."));
     }
 
     @Override
@@ -31,6 +41,9 @@ public class CategoriaService implements IGenericService<Categoria>{
 
     @Override
     public void eliminar(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException("Debe ingresar id");
+        }
         catRepo.deleteById(id);
     }
