+++ src/main/java/com/example/Biblioteca/service/EditorialService.java
@@ -3,8 +3,12 @@ package com.example.Biblioteca.service;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.example.Biblioteca.Excepciones.ValidacionException;
 import com.example.Biblioteca.modelo.Editorial;
 import com.example.Biblioteca.repository.EditorialRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class EditorialService implements IGenericService<Editorial>{
@@ -14,12 +18,19 @@ public class EditorialService implements IGenericService<Editorial>{
 
     @Override
     public void guardar(Editorial t) {
-        editRepo.save(t);
+        if(!editRepo.existsByNombre(t.getNombre())){     // si se agrega otra validación, crear una clase especial para validar editorial
+            editRepo.save(t);
+        }else{
+            throw new ValidacionException("Nombre", "Existe ya un registro de esta editorial");
+        }
     }
 
     @Override
     public Editorial obtenerUno(Long id) {
-       return editRepo.findById(id).orElse(null);
+        if(id == null){
+            throw new IllegalArgumentException("Debe ingresar id");
+        }
+       return editRepo.findById(id).orElseThrow(() -> new EntityNotFoundException("Editorial no encontrado."));
     }
 
     @Override
@@ -29,6 +40,9 @@ public class EditorialService implements IGenericService<Editorial>{
 
     @Override
     public void eliminar(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException("Debe ingresar id");
+        }
         editRepo.deleteById(id);
     }
