+++ src/main/java/com/example/Biblioteca/service/InstitucionService.java
@@ -5,8 +5,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.Biblioteca.Excepciones.ValidacionException;
 import com.example.Biblioteca.modelo.Institucion;
 import com.example.Biblioteca.repository.InstitucionRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class InstitucionService implements IGenericService<Institucion>{
@@ -16,12 +19,19 @@ public class InstitucionService implements IGenericService<Institucion>{
 
     @Override
     public void guardar(Institucion t) {
+        if(!insRepo.existsByNombre(t.getNombre())){     // si se agrega otra validación, crear una clase especial para validar institución
         insRepo.save(t);
+        }else{
+            throw new ValidacionException("Nombre", "Existe ya un registro de esta institución");
+        }
     }
 
     @Override
     public Institucion obtenerUno(Long id) {
-        return insRepo.findById(id).orElse(null);
+        if(id == null){
+            throw new IllegalArgumentException("Debe ingresar id");
+        }
+        return insRepo.findById(id).orElseThrow(() -> new EntityNotFoundException("Institución no encontrada."));
     }
 
     @Override
@@ -31,6 +41,9 @@ public class InstitucionService implements IGenericService<Institucion>{
 
     @Override
     public void eliminar(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException("Debe ingresar id");
+        }
         insRepo.deleteById(id);
     }
